requirements.txt
PyQt5>=5.15
pytesseract>=0.3.10
opencv-python>=4.9.0.80
numpy>=1.24
Pillow>=10.3.0
PyMuPDF>=1.24.0
docx>=0.2.4
python-docx>=1.1.0
paddleocr>=2.7.0 # optional high-accuracy
EasyOCR>=1.7.1   # optional high-accuracy
